requirements.txt
# Core
numpy==1.26.4
scipy==1.16.1
pandas>=2.1
pyarrow>=14
scikit-learn>=1.3
scikit-image==0.25.2
umap-learn>=0.5
opencv-python==4.10.0.84
matplotlib>=3.8
seaborn>=0.13
plotly>=5.20
statsmodels>=0.14
tqdm>=4.67
pyyaml>=6.0
mrcfile==1.5.4
toml==0.10.2
pydantic<2.12

--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.4.0+cu121
torchvision==0.19.0+cu121

# Viz/Notebook
ipykernel
ipywidgets
anywidget
matplotlib
matplotlib-venn==1.1.2

kornia==0.8.0
tensorboard==2.20.0
wandb==0.22.1
PyWavelets==1.9.0
